add api to fetch generation by type and populate type for artifact generations
