+++ legend-depot-artifacts-refresh/src/main/java/org/finos/legend/depot/store/artifacts/services/file/FileGenerationHandlerImpl.java
@@ -15,6 +15,7 @@
 
 package org.finos.legend.depot.store.artifacts.services.file;
 
+import org.apache.commons.io.FilenameUtils;
 import org.finos.legend.depot.artifacts.repository.api.ArtifactRepository;
 import org.finos.legend.depot.artifacts.repository.domain.ArtifactType;
 import org.finos.legend.depot.domain.api.MetadataEventResponse;
@@ -23,7 +24,6 @@ import org.finos.legend.depot.domain.generation.file.StoredFileGeneration;
 import org.finos.legend.depot.domain.project.StoreProjectData;
 import org.finos.legend.depot.domain.version.VersionValidator;
 import org.finos.legend.depot.services.api.generation.file.ManageFileGenerationsService;
-import org.finos.legend.depot.store.artifacts.ArtifactLoadingException;
 import org.finos.legend.depot.store.artifacts.api.generation.file.FileGenerationsArtifactsHandler;
 import org.finos.legend.depot.store.artifacts.api.generation.file.FileGenerationsArtifactsProvider;
 import org.finos.legend.sdlc.domain.model.entity.Entity;
@@ -48,7 +48,7 @@ public class FileGenerationHandlerImpl implements FileGenerationsArtifactsHandle
 {
 
     public static final String TYPE = "type";
-    public static final String PATH = "/";
+    public static final String PATH_SEPARATOR = "/";
     public static final String GENERATION_OUTPUT_PATH = "generationOutputPath";
     public static final String PURE_PACKAGE_SEPARATOR = "::";
     public static final String UNDERSCORE = "_";
@@ -90,7 +90,7 @@ public class FileGenerationHandlerImpl implements FileGenerationsArtifactsHandle
             fileGenerationEntities.forEach(entity ->
             {
                 String generationPath = (String) entity.getContent().get(GENERATION_OUTPUT_PATH);
-                String elementPath = PATH + (generationPath != null ? generationPath : entity.getPath().replace(PURE_PACKAGE_SEPARATOR, UNDERSCORE));
+                String elementPath = PATH_SEPARATOR + (generationPath != null ? generationPath : entity.getPath().replace(PURE_PACKAGE_SEPARATOR, UNDERSCORE));
                 String codeSchemaGenerationType = (String) entity.getContent().get(TYPE);
 
                 generatedFiles.stream().filter(gen -> gen.getPath().startsWith(elementPath)).forEach(gen ->
@@ -108,7 +108,7 @@ public class FileGenerationHandlerImpl implements FileGenerationsArtifactsHandle
             {
                 if (!processedGeneratedFiles.contains(generatedFile))
                 {
-                    Optional<String> entityPath = entityPaths.stream().filter(s -> generatedFile.getPath().startsWith(PATH + s)).findFirst();
+                    Optional<String> entityPath = entityPaths.stream().filter(s -> generatedFile.getPath().startsWith(PATH_SEPARATOR + s)).findFirst();
                     if (!entityPath.isPresent())
                     {
                         String unableToHandle = String.format("Can't find element path for generated file with path %s",generatedFile.getPath());
@@ -138,13 +138,14 @@ public class FileGenerationHandlerImpl implements FileGenerationsArtifactsHandle
 
     private String getExtensionKeyFromGeneration(String path, String entityPath)
     {
-        return Arrays.stream(path.replace(PATH + entityPath + PATH, "").split(PATH)).collect(Collectors.toList()).get(0);
+        String type = Arrays.stream(path.replace(PATH_SEPARATOR + entityPath + PATH_SEPARATOR, "").split(PATH_SEPARATOR)).findFirst().get();
+        return FilenameUtils.getExtension(type).isEmpty() ? type : null;
     }
 
     private Map<String, Entity> buildEntitiesByElementPathMap(List<Entity> entities)
     {
         Map<String, Entity> entityMap = new HashMap<>();
-        entities.forEach(entity -> entityMap.put(entity.getPath().replace(PURE_PACKAGE_SEPARATOR, PATH), entity));
+        entities.forEach(entity -> entityMap.put(entity.getPath().replace(PURE_PACKAGE_SEPARATOR, PATH_SEPARATOR), entity));
         return entityMap;
     }
