+++ legend-depot-artifacts-refresh/src/test/java/org/finos/legend/depot/store/artifacts/services/generation/file/TestGenerationsProvider.java
@@ -161,10 +161,12 @@ public class TestGenerationsProvider extends TestStoreMongo
         Assert.assertFalse(response.hasErrors());
         List<StoredFileGeneration> fileGenerations = generations.getAll();
         Assert.assertNotNull(fileGenerations);
-        Assert.assertEquals(12, fileGenerations.size());
+        Assert.assertEquals(14, fileGenerations.size());
 
         Assert.assertEquals(4, generations.findByType(TEST_GROUP_ID, TEST_ARTIFACT_ID, "2.0.0", "java").size());
         Assert.assertEquals(2, generations.findByType(TEST_GROUP_ID, TEST_ARTIFACT_ID, "2.0.0", "my-ext").size());
+        Assert.assertEquals(2, generations.findByType(TEST_GROUP_ID, TEST_ARTIFACT_ID, "2.0.0", null).size());
+        Assert.assertEquals(0, generations.findByType(TEST_GROUP_ID, TEST_ARTIFACT_ID, "2.0.0", "unkownType").size());
     }
 
     @Test(expected = IllegalArgumentException.class)
